tko.utils) Fix circular reference
